Fix fork requirement for test_multi_connector test since we activate cuda now for MLA
